Add http server mocking
