+++ smart-contract-verifier/smart-contract-verifier-proto/Cargo.toml
@@ -9,11 +9,11 @@ edition = "2021"
 actix-prost = { git = "https://github.com/blockscout/actix-prost" }
 actix-prost-macros = { git = "https://github.com/blockscout/actix-prost" }
 actix-web = "4"
+async-trait = "0.1"
 prost = "0.11"
-tonic = "0.8"
 serde = { version = "1", features = ["derive"] }
 serde_with = { version = "2.0", features = ["hex", "base64"] }
-async-trait = "0.1"
+tonic = "0.8"
 
 anyhow = { version = "1.0.76", optional = true }
 reqwest = { version = "0.11", optional = true }
@@ -21,15 +21,13 @@ reqwest-middleware = { version = "0.2", optional = true }
 reqwest-retry = { version = "0.3", optional = true }
 thiserror = { version = "1", optional = true }
 url = { version = "2.5.0", optional = true }
-wiremock = { version = "0.5.22", optional = true }
+mockall = { version = "0.12.1", optional = true }
+tokio = { version = "1", optional = true }
 
 [build-dependencies]
 actix-prost-build = { git = "https://github.com/blockscout/actix-prost" }
 tonic-build = "0.8"
 prost-build = "0.11"
-
-[dev-dependencies]
-tokio = {version = "1.35.1", features = ["full"]}
 
 [features]
 http-client = [
@@ -41,5 +39,6 @@ http-client = [
     "dep:url",
 ]
 mock = [
-    "dep:wiremock",
+    "dep:mockall",
+    "dep:tokio",
 ]
